add test functions for 3 simple rules
