+++ imports/api/node_rules/register-server.js
@@ -1,4 +1,4 @@
-import './functions';
+import './index';
 import './SLAs/B2B/5';
 import './SLAs/B2B/6';
 import './SLAs/B2B/7';
